Fix SP permission check failing due to _metadata import and OBO ACL filtering
The auto-optimize permission check always reported "Missing permissions"
because importing from genie_space_optimizer.backend.routes.settings
triggered a transitive import of _metadata.py (generated during GSO
standalone build, not present in the main app), causing ModuleNotFoundError.

Extract SP permission utilities (get_sp_principal_aliases,
probe_sp_required_access) into a new lightweight common/sp_permissions.py
module with no GSO framework core dependency. Update auto_optimize.py to
import from there instead of settings.py.

Also fix sp_can_manage_space() to try the SP client even when the OBO
client returns a valid ACL response, since OBO tokens may return a
filtered view that omits the SP's own ACL entry.
